# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Student-School Portal API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/teachers/class": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "View the class assigned to the authenticated teacher",
                "operationId": "c9c08494d675e3538cabc31216ef8042",
                "responses": {
                    "200": {
                        "description": "Class details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "5th Grade"
                                        },
                                        "school_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "teacher_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No class assigned to this teacher"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/teachers/class/students": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "View the students in the class assigned to the authenticated teacher",
                "operationId": "93b9c16c77fecb3e4a952997090dd63c",
                "responses": {
                    "200": {
                        "description": "List of students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "johndoe@example.com"
                                            },
                                            "class_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No class assigned to this teacher"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}